Refactor mask tab
